moved drag logic to app and refactored code
